Put all ItemRecyclerView related classes in the ItemAdapter
