+++ app/src/main/java/com/example/sharingang/items/ItemsAdapter.kt
@@ -1,10 +1,14 @@
 package com.example.sharingang.items
 
+import android.view.LayoutInflater
 import android.view.ViewGroup
+import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
+import androidx.recyclerview.widget.RecyclerView
+import com.example.sharingang.databinding.ItemViewBinding
 
 class ItemsAdapter(private val clickListener: ItemListener) :
-    ListAdapter<Item, ItemViewHolder>(ItemsDiffCallback()) {
+    ListAdapter<Item, ItemsAdapter.ItemViewHolder>(ItemsDiffCallback()) {
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ItemViewHolder {
         return ItemViewHolder.from(parent)
@@ -13,4 +17,40 @@ class ItemsAdapter(private val clickListener: ItemListener) :
     override fun onBindViewHolder(holder: ItemViewHolder, position: Int) {
         holder.bind(getItem(position)!!, clickListener)
     }
+
+    class ItemViewHolder private constructor(private val binding: ItemViewBinding) :
+        RecyclerView.ViewHolder(binding.root) {
+
+        fun bind(item: Item, clickListener: ItemListener) {
+            binding.itemListViewTitle.text = item.description
+            binding.item = item
+            binding.clickListener = clickListener
+        }
+
+        companion object {
+            fun from(parent: ViewGroup): ItemViewHolder {
+                val layoutInflater = LayoutInflater.from(parent.context)
+                val binding = ItemViewBinding.inflate(layoutInflater, parent, false)
+                return ItemViewHolder(binding)
+            }
+        }
+    }
+}
+
+class ItemsDiffCallback : DiffUtil.ItemCallback<Item>() {
+    override fun areItemsTheSame(oldItem: Item, newItem: Item): Boolean {
+        // Since it's a data class, the == checks if all the fields are equal
+        // now there is only a description so it might work weirdly with same description items,
+        // but later we'll check if their IDs are equal or not
+        return oldItem == newItem
+    }
+
+    override fun areContentsTheSame(oldItem: Item, newItem: Item): Boolean {
+        return oldItem == newItem
+    }
+}
+
+// Replace the description with its ID once the database is set
+class ItemListener(val clickListener: (item: Item) -> Unit) {
+    fun onClick(item: Item) = clickListener(item)
 }
